Unavailable items are now hidden by the API
If Mega provides us with items that we do not have a key for or items
that have invalid data (e.g. a file without a size), the API just hides
this item and continues as if it did not exist.

This greatly simplifies API usage, since you never really care about
such weird cases, you only want to focus on the positive scenario.