Added filesystem tests; fixed invalidation
Filesystem invalidation previously depended on the server notifying us
of a change; because this is an asynchronous process, the client would
return a stale filesystem snapshot if you asked for one immediately
after performing an operation. Now, it invalidates the snapshot after
every operation.

Added some more filesystem tests - also found another bug, that moving
items is not working properly.